requirements.txt
fastapi
firebase-admin
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit>=1.37